Split chrome-scripts/classes into dom/ and utils/
The classes/ folder had no classes left after the function-style
migration. Move DOM helpers (createElement, createObserver,
createPercentElement) to dom/, and the HTTP wrapper makeRequest
into utils/ alongside the other low-level helpers.
