+++ src/chrome-scripts/dom/createPercentElement.ts
