+++ src/chrome-scripts/dom/index.ts
@@ -1,6 +1,5 @@
 import { createElement } from './createElement';
 import { createPercentElement } from './createPercentElement';
 import { createObserver } from './createObserver';
-import { makeRequest } from './makeRequest';
 
-export { createElement, createPercentElement, createObserver, makeRequest };
+export { createElement, createPercentElement, createObserver };
