+++ src/chrome-scripts/utils/index.ts
@@ -1,5 +1,6 @@
 import { debounce } from './debounce';
 import { logger } from './logger';
 import { badge } from './badge';
+import { makeRequest } from './makeRequest';
 
-export { debounce, logger, badge };
+export { debounce, logger, badge, makeRequest };
